Feat: 비밀번호 수정 ui 이동, 에러 문구 추가, eventListener 관련 함수 파일 분리
